pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crits"
version = "5.0.0-dev"
description = "Collaborative Research Into Threats - Threat Intelligence Platform"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "MITRE Corporation" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
]
dependencies = [
    # Web Framework - Django for legacy UI
    "Django>=4.2,<5.0",

    # GraphQL API - FastAPI + Strawberry (runs alongside Django)
    "fastapi>=0.109.0",
    "strawberry-graphql[fastapi]>=0.220.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic-settings>=2.1.0",

    # MongoDB
    "mongoengine>=0.28.0",
    "pymongo>=4.6.0",
    "django-mongoengine>=0.5.0",

    # Task Queue & Cache
    "celery>=5.3.0",
    "redis>=5.0.0",

    # NOTE: django-tastypie removed - replaced by Strawberry GraphQL

    # Security & Cryptography
    "cryptography>=42.0.0",
    "pycryptodome>=3.20.0",  # TOTP support (AES encryption)
    "python-magic>=0.4.27",

    # XML Processing
    "lxml>=5.0.0",
    "defusedxml>=0.7.0",

    # Image Processing
    "Pillow>=10.0.0",

    # File Format Parsing
    "biplist>=1.0.3",
    "olefile>=0.47",
    "chardet>=5.0.0",

    # HTTP & Networking
    "requests>=2.31.0",

    # WSGI Server
    "gunicorn>=21.0.0",
    "whitenoise>=6.6.0",  # Static file serving for production

    # Date/Time & Parsing
    "python-dateutil>=2.8.0",
    "pyparsing>=3.1.0",
    "pytz>=2024.1",

    # Version parsing (replaces deprecated distutils.version)
    "packaging>=23.0",

    # Configuration
    "PyYAML>=6.0.0",

    # TOTP/2FA
    "qrcode>=7.4.0",
]

[project.optional-dependencies]
ldap = [
    "python-ldap>=3.4.0",
]
# Fuzzy hashing requires system library: apt-get install libfuzzy-dev
fuzzy = [
    "pydeep>=0.4",
    "pyimpfuzzy>=0.5",
]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.8.0",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "django-debug-toolbar>=4.2.0",
    "ipython>=8.0.0",
    # Type stubs
    "types-redis",
    "types-PyYAML",
    "types-python-dateutil",
]

[project.urls]
Homepage = "https://github.com/mitre/crits"
Documentation = "https://github.com/mitre/crits/wiki"
Repository = "https://github.com/mitre/crits"

[tool.hatch.build.targets.wheel]
packages = ["crits"]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "migrations",
    "node_modules",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument
    "SIM108", # ternary operator
]

[tool.ruff.lint.isort]
known-first-party = ["crits"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
strict_equality = true
show_error_codes = true
show_column_numbers = true

# Default: ignore missing imports for legacy code
ignore_missing_imports = true

# Strict mode for crits_api (new code)
[[tool.mypy.overrides]]
module = "crits_api.*"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
# Allow Any returns from MongoEngine querysets
warn_return_any = false

# Ignore legacy crits module (too many issues to fix)
[[tool.mypy.overrides]]
module = "crits.*"
ignore_errors = true

# Third-party modules without stubs
[[tool.mypy.overrides]]
module = [
    "mongoengine.*",
    "celery.*",
    "PIL.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "crits.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"

[tool.uv.extra-build-dependencies]
django-mongoengine = ["setuptools<78"]
biplist = ["setuptools<78"]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.8.0",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "django-debug-toolbar>=4.2.0",
    "ipython>=8.0.0",
    # Type stubs
    "types-redis",
    "types-PyYAML",
    "types-python-dateutil",
]

# Note: Optional fuzzy hashing (pydeep, pyimpfuzzy) requires ssdeep library:
#   Ubuntu/Debian: apt-get install libfuzzy-dev
#   macOS: brew install ssdeep
#   Then install: uv sync --extra fuzzy
